requirements.txt
mcp>=1.0.0
httpx>=0.25.0
pydantic>=2.0.0
python-dotenv>=1.0.0
uvicorn>=0.24.0
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0 